refactor: 페이지 종속 컴포넌트는 페이지 폴더 하위로 이동
