--- frontend/src/components/Detail/index.ts
@@ -1,2 +1,0 @@
-export { default as DetailContent } from './Content';
-export { default as DetailSideBar } from './SideBar';
