+++ frontend/src/components/GroupEdit/index.tsx
@@ -5,7 +5,6 @@ import { useRecoilValue, useResetRecoilState } from 'recoil';
 
 import { requestEditGroup } from 'apis/request/group';
 import Modal from 'components/@shared/Modal';
-import CalendarEditor from 'components/Create/Step5/CalendarEditor';
 import { QUERY_KEY } from 'constants/key';
 import { GUIDE_MESSAGE } from 'constants/message';
 import { GROUP_RULE } from 'constants/rule';
@@ -13,6 +12,8 @@ import useClosingState from 'hooks/useClosingState';
 import useCreateState from 'hooks/useCreateState';
 import useModal from 'hooks/useModal';
 import useSnackbar from 'hooks/useSnackbar';
+// TODO: 수정
+import CalendarEditor from 'pages/Create/Steps/Step5/CalendarEditor';
 import validateGroupData from 'pages/Create/validate';
 import { categoryState, groupDetailState, modalState } from 'store/states';
 import { CategoryType } from 'types/data';
