+++ frontend/src/components/SearchForm/index.styled.ts
