+++ frontend/src/pages/Create/Navigator/index.styled.ts
