+++ frontend/src/pages/Create/Steps/@shared/styled.ts
