+++ frontend/src/pages/Create/Steps/Step2/index.styled.ts
