+++ frontend/src/pages/Create/Steps/Step4/index.styled.ts
