+++ frontend/src/pages/Create/Steps/Step5/index.styled.ts
