+++ frontend/src/pages/Create/Steps/Step6/index.tsx
