+++ frontend/src/pages/Create/Steps/Step7/index.styled.ts
