+++ frontend/src/pages/Create/Steps/Step8/index.styled.ts
