+++ frontend/src/pages/Detail/Content/index.styled.ts
