+++ frontend/src/pages/Detail/Content/index.tsx
