+++ frontend/src/pages/Detail/SideBar/Info/ControlButton/index.styled.ts
