+++ frontend/src/pages/Detail/SideBar/Info/index.styled.ts
