+++ frontend/src/pages/Detail/SideBar/Participants/index.styled.ts
