+++ frontend/src/pages/Detail/SideBar/index.styled.ts
