+++ frontend/src/pages/Detail/SideBar/index.tsx
