+++ frontend/src/pages/Main/Category/index.styled.ts
