+++ frontend/src/pages/Main/RecommendGroups/index.styled.ts
