+++ frontend/src/pages/Main/SearchSection/index.styled.ts
