+++ frontend/src/pages/MyGroup/JoinedGroups/index.tsx
