+++ frontend/src/pages/MyInformation/Withdrawal/index.styled.ts
